requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
ffmpeg-python==0.2.0
frozenlist==1.6.0
future==1.0.0
idna==3.10
imageio-ffmpeg==0.6.0
makefun==1.16.0
multidict==6.4.3
numpy==2.2.5
propcache==0.3.1
py-cord==2.6.1
python-dotenv==1.1.0
yarl==1.20.0
